Rewrote timer code.
The new code is based on timestamps, rather than delays. This eliminates the
need to update every timer in timer_process(), which cost a non-trivial amount
of CPU time on low-end machines.

This involved changes to every user of the timer code, so this change almost
certainly introduces bugs.

+++ src/cpu.h
@@ -140,7 +140,7 @@ void cpu_WRMSR();
 
 extern int cpu_use_dynarec;
 
-extern int xt_cpu_multi;
+extern uint64_t xt_cpu_multi;
 
 extern int isa_cycles;
 #define ISA_CYCLES(x) (x * isa_cycles)
